backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.44
asyncpg==0.31.0
pydantic==2.12.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
boto3==1.42.28
strands-agents==1.27.0
strands-agents-tools==0.2.21
pymupdf==1.27.1
pytesseract==0.3.13
httpx==0.28.1
psycopg2-binary==2.9.11
python-socketio==5.11.2
numpy==1.26.4